Fixed a bug where the combobox was not displaying the selected passenger. Implemented a system where the selected passengers seat will turn green. Moved part of the functionality of UpdateFlightDetails to ResetSeats to avoid rebinding the choose passenger combo box unnecessarily. Added a confirmation message to the delete passenger function.

+++ Assignment6AirlineReservation/MainWindow.xaml.cs
@@ -24,15 +24,30 @@ namespace Assignment6AirlineReservation
     /// </summary>
     public partial class MainWindow : Window
     {
-        //clsDataAccess clsData;
+        /// <summary>
+        /// The Add Passenger window
+        /// </summary>
         wndAddPassenger wndAddPass;
 
+        /// <summary>
+        /// Flight Manager
+        /// </summary>
         clsFlightManager flightManager;
+
+        /// <summary>
+        /// Passenger Manager
+        /// </summary>
         clsPassengerManager passengerManager;
 
-        // Determines if the program is in "Add Passenger Mode"
+        /// <summary>
+        /// Determines if the program is in "Add Passenger Mode"
+        /// </summary>
         public static bool bAddingPassenger = false;
-        public bool bChangingSeat = false; // I don't think this one needs to be static
+
+        /// <summary>
+        /// Determines if the program is in "Change Seat Mode"
+        /// </summary>
+        public bool bChangingSeat = false;
 
         public MainWindow()
         {
@@ -88,6 +103,7 @@ namespace Assignment6AirlineReservation
         {
             try
             {
+                // Initialize passenger manager
                 passengerManager = new clsPassengerManager();
 
                 // This list holds all passengers on the current flight
@@ -96,74 +112,78 @@ namespace Assignment6AirlineReservation
                 // Bind passengers to the passenger combo box
                 cbChoosePassenger.ItemsSource = passengers;
 
-
-                if (flightID == 1)
-                {
-                    // Display appropriate flight 
-                    CanvasA380.Visibility = Visibility.Hidden;
-                    Canvas767.Visibility = Visibility.Visible;
-
-                    // Mark taken seats red and avaliable seats blue
-                    foreach (Label seat in c767_Seats.Children) 
-                    {
-                        // Reset each seat to blue once
-                        seat.Background = new SolidColorBrush(Colors.Blue);
-
-                        // Iterate through passengers
-                        foreach (clsPassenger passenger in passengers)
-                        {
-                            //TODO: This if won't ever be true. Remove
-                            //if (seat.Background == new SolidColorBrush(Colors.Red))
-                            //{
-                            //    break;
-                            //}
-
-                            // The seat is taken
-                            if (seat.Name == ("Seat" + passenger.seatNumber))
-                            {
-                                seat.Background = new SolidColorBrush(Colors.Red);
-                                break;
-                            }
-                        }
-                    }
-                }
-                else 
-                {
-                    // Display appropriate flight
-                    Canvas767.Visibility = Visibility.Hidden;
-                    CanvasA380.Visibility = Visibility.Visible;
-
-                    // Mark taken seats red and avaliable seats blue
-                    foreach (Label seat in cA380_Seats.Children)
-                    {
-                        // Reset each seat to blue once
-                        seat.Background = new SolidColorBrush(Colors.Blue);
-
-                        // Iterate through passengers
-                        foreach (clsPassenger passenger in passengers)
-                        {
-                            //TODO: This if won't ever be true. Remove
-                            if (seat.Background == new SolidColorBrush(Colors.Red))
-                            {
-                                break;
-                            }
-
-                            // The seat is taken
-                            if (seat.Name == ("SeatA" + passenger.seatNumber))
-                            {
-                                seat.Background = new SolidColorBrush(Colors.Red);
-                                break;
-                            }
-                            
-                        }
-                    }
-                }
+                // Reset the color of all seats to blue and mark taken seats red
+                ResetSeats(flightID, ref passengers);
             }
             catch (Exception ex)
             {
                 //Throw an exception
                 throw new Exception(MethodInfo.GetCurrentMethod().DeclaringType.Name + "." +
                                     MethodInfo.GetCurrentMethod().Name + " -> " + ex.Message);
+            }
+        }
+
+        /// <summary>
+        /// Resets all seats to blue and then marks taken seats as red
+        /// </summary>
+        /// <param name="flightID">Flight ID of the current flight</param>
+        /// <param name="passengers">Reference to a list of passengers</param>
+        public void ResetSeats(int flightID, ref List<clsPassenger> passengers)
+        {
+            if (flightID == 1)
+            {
+                // Display appropriate flight 
+                CanvasA380.Visibility = Visibility.Hidden;
+                Canvas767.Visibility = Visibility.Visible;
+
+                // Mark taken seats red and avaliable seats blue
+                foreach (Label seat in c767_Seats.Children)
+                {
+                    // Reset each seat to blue once
+                    seat.Background = new SolidColorBrush(Colors.Blue);
+
+                    // Iterate through passengers
+                    foreach (clsPassenger passenger in passengers)
+                    {
+                        // The seat is taken
+                        if (seat.Name == ("Seat" + passenger.seatNumber))
+                        {
+                            seat.Background = new SolidColorBrush(Colors.Red);
+                            break;
+                        }
+                    }
+                }
+            }
+            else
+            {
+                // Display appropriate flight
+                Canvas767.Visibility = Visibility.Hidden;
+                CanvasA380.Visibility = Visibility.Visible;
+
+                // Mark taken seats red and avaliable seats blue
+                foreach (Label seat in cA380_Seats.Children)
+                {
+                    // Reset each seat to blue once
+                    seat.Background = new SolidColorBrush(Colors.Blue);
+
+                    // Iterate through passengers
+                    foreach (clsPassenger passenger in passengers)
+                    {
+                        //TODO: This if won't ever be true. Remove
+                        if (seat.Background == new SolidColorBrush(Colors.Red))
+                        {
+                            break;
+                        }
+
+                        // The seat is taken
+                        if (seat.Name == ("SeatA" + passenger.seatNumber))
+                        {
+                            seat.Background = new SolidColorBrush(Colors.Red);
+                            break;
+                        }
+
+                    }
+                }
             }
         }
 
@@ -179,6 +199,8 @@ namespace Assignment6AirlineReservation
             {
                 // Initialize add passenger window
                 wndAddPass = new wndAddPassenger();
+
+                // Display the add passenger window
                 wndAddPass.ShowDialog();
 
                 // If not in add passenger mode, don't disable controls
@@ -187,7 +209,7 @@ namespace Assignment6AirlineReservation
                     return;
                 }
 
-                // Disable all controls except choosing a seat
+                // Disable all controls except those for choosing a seat
                 cbChooseFlight.IsEnabled = false;
                 cbChoosePassenger.IsEnabled = false;
                 cmdAddPassenger.IsEnabled = false;
@@ -202,9 +224,12 @@ namespace Assignment6AirlineReservation
             }
         }
 
-
-
-
+        /// <summary>
+        /// If in normal mode it will display who is assigned to a seat if someone is.
+        /// If in add passenger or change seat mode the selected seat will be assigned to the passenger.
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void Seat_Click(object sender, MouseButtonEventArgs e)
         {
             try
@@ -227,15 +252,21 @@ namespace Assignment6AirlineReservation
                 // Isolate the value (seat number) of seatMatch
                 string sSeatNumber = seatMatch.Value;
 
-                // Seat is already taken select the passenger belonging to the clicked seat
+                // Seat is already taken, select the passenger belonging to the clicked seat
                 if (flightManager.SeatTaken(sSeatNumber, flightID))
                 {
+                    // Iterate through passengers
                     foreach (clsPassenger passenger in cbChoosePassenger.ItemsSource)
                     {
                         if (sSeatNumber == passenger.seatNumber)
                         {
+                            // Select chosen passenger in combo box
                             cbChoosePassenger.SelectedItem = passenger;
+                            
+                            // Display passengers seat number in seat number label
                             lblPassengersSeatNumber.Content = passenger.seatNumber;
+
+                            // Do nothing else
                             return;
                         }
                     }
@@ -306,6 +337,7 @@ namespace Assignment6AirlineReservation
                     return;
                 }
 
+                // Enter seat changing mode
                 bChangingSeat = true;
 
                 // Disable all controls except choosing a seat
@@ -342,22 +374,30 @@ namespace Assignment6AirlineReservation
             // Get the flight ID
             int flightID = cbChooseFlight.SelectedIndex + 1;
 
-            // Delete the passenger from the database
-            passengerManager.DeletePassenger(passenger, flightID);
+            // A confirmation message string
+            string sConfirmationMessage = $"Are you sure you would like to delete {passenger.firstName} {passenger.lastName} " +
+                $"from flight {flightID}? This action cannot be undone.";
 
-            UpdateFlightDetails(flightID);
+            // Stores the result of the confirmation message
+            MessageBoxResult result = MessageBox.Show
+                (sConfirmationMessage, "Delete Passenger", MessageBoxButton.YesNo, MessageBoxImage.Warning);
+
+            if (result == MessageBoxResult.Yes)
+            {
+                // Delete the passenger from the database
+                passengerManager.DeletePassenger(passenger, flightID);
+
+                UpdateFlightDetails(flightID);
+            }
         }
 
         /// <summary>
-        /// Highlights the selected passengers seat seat
+        /// Highlights the selected passengers seat
         /// </summary>
         /// <param name="sender"></param>
         /// <param name="e"></param>
         private void cbChoosePassenger_SelectionChanged(object sender, SelectionChangedEventArgs e)
         {
-            // TODO: cbChoosePassenger combo box does not seem to be showing the selected passenger
-            // This may have something to do with this fuction
-
             // No passenger is selected. Do nothing.
             if (cbChoosePassenger.SelectedItem == null || cbChoosePassenger.SelectedIndex == -1)
             {
@@ -367,11 +407,14 @@ namespace Assignment6AirlineReservation
             // ID of the current flight
             int flightID = cbChooseFlight.SelectedIndex + 1;
 
-            // Represents the currently selected passenger
+            // All passengers on the current flight
+            List<clsPassenger> passengers = passengerManager.GetPassengers(flightID);
+
+            // The currently selected passenger
             clsPassenger selectedPassenger = (clsPassenger)cbChoosePassenger.SelectedItem;
 
             // Reset flight details
-            UpdateFlightDetails(flightID);
+            ResetSeats(flightID, ref passengers);
 
             // Set the selected passengers seat to selected (green)
             UpdateFlightDetails(flightID, selectedPassenger.seatNumber);
